move command_matcher into separate file
